#659 Allocate memory from bottom to top
