+++ evm_loader/program/src/entrypoint.rs
@@ -2,17 +2,11 @@
 
 #![cfg(not(feature = "no-entrypoint"))]
 
-use std::{
-    alloc::Layout,
-    mem::size_of,
-    ptr::null_mut,
-    usize
-};
 
 use solana_program::{
     account_info::AccountInfo,
     entrypoint,
-    entrypoint::{ProgramResult, HEAP_START_ADDRESS},
+    entrypoint::{ProgramResult},
     program_error::ProgramError,
     pubkey::Pubkey,
 };
@@ -23,59 +17,6 @@ use crate::{
 };
 
 
-const HEAP_LENGTH: usize = 256*1024;
-
-/// Developers can implement their own heap by defining their own
-/// `#[global_allocator]`.  The following implements a dummy for test purposes
-/// but can be flushed out with whatever the developer sees fit.
-pub struct BumpAllocator;
-
-impl BumpAllocator {
-    /// Get occupied memory
-    #[inline]
-    #[must_use]
-    #[allow(clippy::missing_const_for_fn)]
-    #[allow(clippy::pedantic)]
-    pub fn occupied() -> usize {
-        const POS_PTR: *mut usize = HEAP_START_ADDRESS as *mut usize;
-        const TOP_ADDRESS: usize = HEAP_START_ADDRESS as usize + HEAP_LENGTH;
-
-        let pos = unsafe{*POS_PTR};
-        if pos == 0 {0} else {TOP_ADDRESS-pos}
-    }
-}
-
-unsafe impl std::alloc::GlobalAlloc for BumpAllocator {
-    #[inline]
-    #[allow(clippy::pedantic)]
-    unsafe fn alloc(&self, layout: Layout) -> *mut u8 {
-        const POS_PTR: *mut usize = HEAP_START_ADDRESS as *mut usize;
-        const TOP_ADDRESS: usize = HEAP_START_ADDRESS as usize + HEAP_LENGTH;
-        const BOTTOM_ADDRESS: usize = HEAP_START_ADDRESS as usize + size_of::<*mut u8>();
-
-        let mut pos = *POS_PTR;
-        if pos == 0 {
-            // First time, set starting position
-            pos = TOP_ADDRESS;
-        }
-        pos = pos.saturating_sub(layout.size());
-        pos &= !(layout.align().saturating_sub(1));
-        if pos < BOTTOM_ADDRESS {
-            return null_mut();
-        }
-
-        *POS_PTR = pos;
-        pos as *mut u8
-    }
-    #[inline]
-    unsafe fn dealloc(&self, _: *mut u8, _layout: Layout) {
-        // I'm a bump allocator, I don't free
-    }
-}
-
-#[cfg(target_arch = "bpf")]
-#[global_allocator]
-static mut A: BumpAllocator = BumpAllocator;
 
 entrypoint!(process_instruction);
 
@@ -140,6 +81,6 @@ fn process_instruction<'a>(
         _ => Err!(ProgramError::InvalidInstructionData; "Invalid instruction"),
     };
 
-    solana_program::msg!("Total memory occupied: {}", &BumpAllocator::occupied());
+    solana_program::msg!("Total memory occupied: {}", crate::allocator::BumpAllocator::occupied());
     result
 }
