+++ evm_loader/program/src/lib.rs
@@ -7,10 +7,11 @@
 #![allow(missing_docs, clippy::missing_panics_doc, clippy::missing_errors_doc)]
 
 
+mod allocator;
 #[macro_use]
 mod debug;
 #[macro_use]
-mod error;
+pub mod error;
 pub mod account;
 pub mod account_storage;
 pub mod config;
